# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: API de recordatorios de medicamentos
  version: 0.3.0
  x-closed-phase: 2
  description: |
    Contrato de la API. Fuente de verdad del backend y de los clientes generados por el frontend
    (Orval para la web, openapi-generator para el reloj Wear OS).

    Convenciones:
    - Rutas en kebab-case y plural; campos JSON en camelCase; fechas ISO 8601 con zona; IDs UUID.
    - Errores en RFC 9457 (`application/problem+json`) con `type`, `title`, `status`, `detail`, `code`.
    - Paginación por cursor (`?cursor=&limit=`) con respuesta `{ items, nextCursor }`.

    Autenticación (`Authorization: Bearer <accessToken>`, JWT de 15 min). Hay dos tipos de sesión:
    - **Usuario (web)**: se abre con `POST /auth/register` o `POST /auth/login` y se renueva con
      `POST /auth/refresh`. El token de acceso va en el JSON; el **refresh token nunca va en el
      JSON**: viaja solo en la cookie `__Secure-refresh-token` (`HttpOnly; Secure;
      SameSite=Strict; Path=/auth; Max-Age=2592000`). El navegador la envía sola a `/auth/*`; la
      web debe llamar con `credentials: "include"`. CORS con credenciales solo para el origen del
      panel web. `POST /auth/logout` la borra (`Max-Age=0`). Su token de acceso no sirve en
      `/devices/me/*` (403).
    - **Reloj**: se abre con `POST /devices/token` (`grantType: device_code`) y se renueva con el
      mismo endpoint (`grantType: refresh_token`). Recibe **ambos tokens en JSON** (no usa
...
    revoca sus tokens.

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /devices/me/dose-events:
    post:
      tags: [doses]
      operationId: createMyDoseEvents
      x-phase: 2
      summary: Subir eventos de toma (lote, idempotente)
      description: |
        Solo con token de reloj. Hasta 100 eventos por lote. Idempotente por `eventId`: reenviar un
        evento ya guardado devuelve `outcome: duplicate` sin duplicar nada. Cada evento se
        resuelve por separado y el lote responde 200; un evento inválido no hace fallar a los
        demás. Un evento solo se acepta si `scheduleId` pertenece al paciente del reloj y
        `scheduledAt` es una hora que ese horario generaba; si no, `outcome: rejected` con
        `code` (`schedule_not_found`, `invalid_scheduled_at`). Una dosis tiene como mucho un
        evento: si llega otro `eventId` para la misma dosis (`scheduleId` + `scheduledAt`), gana
        el primero y el segundo se rechaza con `code: dose_already_recorded`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoseEventBatch"
      responses:
        "200":
          description: Resultado por evento, en el mismo orden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoseEventBatchResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /patients/{patientId}/dose-history:
    parameters:
      - $ref: "#/components/parameters/PatientIdPath"
    get:
      tags: [doses]
      operationId: listDoseHistory
      x-phase: 2
      summary: Historial de dosis de un paciente
      description: |
        Dosis ya vencidas, de la más reciente a la más antigua, con su estado. Una dosis sin
        evento cuya hora programada pasó hace más de 60 minutos se informa como `MISSED`
        (se calcula al consultar). Las dosis sin evento futuras o dentro de la ventana de 60 min
        no aparecen; una dosis con evento aparece siempre. Solo cuentan las dosis que el horario
        ya generaba cuando se creó o editó, y mientras el medicamento no estaba archivado.
        Rango por defecto: últimos 7 días; máximo 90 días, y `from` no puede ser posterior a
        `to` (422).
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
      responses:
        "200":
          description: Página del historial.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoseHistoryPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /patients/{patientId}/adherence:
    parameters:
      - $ref: "#/components/parameters/PatientIdPath"
    get:
      tags: [doses]
      operationId: getAdherence
      x-phase: 2
      summary: Porcentaje de adherencia de un paciente
      description: |
        `percentage = taken / (taken + skipped + missed) * 100` sobre las dosis vencidas del
        rango (mismas reglas y rango por defecto que el historial). `null` si no hubo dosis
        vencidas.
      parameters:
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
      responses:
        "200":
          description: Adherencia del rango.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adherence"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de acceso de 15 min. Dos tipos: de usuario (web) y de reloj. `/devices/me` y
        `/devices/me/*` solo aceptan tokens de reloj; todas las demás rutas protegidas, solo
        tokens de usuario. El tipo equivocado responde 403 (`code: wrong_token_type`).
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Cursor opaco devuelto como `nextCursor` en la página anterior.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PatientIdPath:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    FromQuery:
      name: from
      in: query
      required: false
      description: Inicio del rango (inclusive), fecha local del paciente.
      schema:
        type: string
        format: date
    ToQuery:
      name: to
      in: query
      required: false
      description: Fin del rango (inclusive), fecha local del paciente.
      schema:
        type: string
        format: date
...
  responses:
    Unauthorized:
      description: Falta el token, es inválido o caducó (o credenciales incorrectas en login).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: |
        Autenticado pero sin permiso: rol que no permite la acción sobre un recurso accesible, o
        tipo de token equivocado (`wrong_token_type`). Nunca se usa para pacientes sin vínculo
        (eso es 404).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: |
        No existe o el usuario no tiene acceso: un cuidador que pide un paciente sin vínculo
        activo (o sus recursos) recibe 404, igual que si no existiera, para no revelar datos
        ajenos.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ValidationError:
      description: La petición no cumple el esquema.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"
...
  schemas:
...
    Problem:
      type: object
      description: Error según RFC 9457.
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
          description: Código estable legible por máquina (snake_case).
        instance:
          type: string
          format: uri-reference
      additionalProperties: true
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              items:
                type: object
                required: [loc, msg, type]
                properties:
                  loc:
                    type: array
                    items:
                      oneOf:
                        - type: string
                        - type: integer
                  msg:
                    type: string
                  type:
                    type: string
...
    DoseEventStatus:
      type: string
      enum: [TAKEN, SKIPPED]
    DoseEventInput:
      type: object
      required: [eventId, scheduleId, scheduledAt, status, actedAt]
      properties:
        eventId:
          type: string
          format: uuid
          description: Generado por el reloj; clave de idempotencia.
        scheduleId:
          type: string
          format: uuid
        scheduledAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/DoseEventStatus"
        actedAt:
          type: string
          format: date-time
          description: Cuándo pulsó el paciente en el reloj.
    DoseEventBatch:
      type: object
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/DoseEventInput"
    DoseEventResult:
      type: object
      required: [eventId, outcome]
      properties:
        eventId:
          type: string
          format: uuid
        outcome:
          type: string
          enum: [created, duplicate, rejected]
        code:
          type: [string, "null"]
          description: Presente solo si `rejected`.
    DoseEventBatchResult:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/DoseEventResult"
    DoseStatus:
      type: string
      enum: [TAKEN, SKIPPED, MISSED]
    DoseHistoryItem:
      type: object
      required: [scheduleId, medicationId, medicationName, dosage, scheduledAt, status, actedAt]
      properties:
        scheduleId:
          type: string
          format: uuid
        medicationId:
          type: string
          format: uuid
        medicationName:
          type: string
        dosage:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/DoseStatus"
        actedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
          description: "`null` si `MISSED`."
    DoseHistoryPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DoseHistoryItem"
        nextCursor:
          oneOf:
            - type: string
            - type: "null"
    Adherence:
      type: object
      required: [from, to, taken, skipped, missed, percentage]
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        taken:
          type: integer
          minimum: 0
        skipped:
          type: integer
          minimum: 0
        missed:
          type: integer
          minimum: 0
        percentage:
          oneOf:
            - type: number
              minimum: 0
              maximum: 100
            - type: "null"